# docs/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kumbukum API",
    "version": "1.0.0",
    "description": "Everything your AI needs to remember — Notes, Memory, URLs, AI Chat"
  },
  "servers": [
    {
      "url": "https://your-instance.example.com/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from login"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Personal access token. Value: `Token <your-token>`"
      }
    },
    "schemas": {
...
      "Note": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "host_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Memory": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "host_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Url": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "screenshot": {
            "type": "string",
            "description": "Screenshot filename (hash)"
          },
          "screenshot_url": {
            "type": "string",
            "description": "Signed URL to screenshot image (time-limited)"
          },
          "project": {
            "type": "string"
          },
          "crawl_enabled": {
            "type": "boolean"
          },
          "host_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Email": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "message_id": {
            "type": "string"
          },
          "references": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "in_reply_to": {
            "type": "string"
          },
          "from": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "string"
          },
          "text_content": {
            "type": "string"
          },
          "attachment_text_content": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "api",
              "emailforwarding"
            ]
          },
          "project": {
            "type": "string"
          },
          "host_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ByoAiProviderStatus": {
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "masked": {
            "type": "string",
            "description": "Masked placeholder when a custom key is configured"
          }
        }
      },
      "ByoAiScopeSettings": {
        "type": "object",
        "properties": {
          "openai_api_key": {
            "$ref": "#/components/schemas/ByoAiProviderStatus"
          },
          "gemini_api_key": {
            "$ref": "#/components/schemas/ByoAiProviderStatus"
          }
        }
      },
      "ByoAiScopeUpdate": {
        "type": "object",
        "properties": {
          "openai_api_key": {
            "type": "string",
            "nullable": true,
            "description": "New OpenAI API key. Use null or empty string to clear."
          },
          "gemini_api_key": {
            "type": "string",
            "nullable": true,
            "description": "New Gemini API key. Use null or empty string to clear."
          }
        }
      },
...
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 50
        }
      },
      "project": {
        "name": "project",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Filter by project ID"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/notes": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "List notes",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/project"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Note"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "summary": "Create a note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "project": {
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "note": {
                      "$ref": "#/components/schemas/Note"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/memories": {
      "get": {
        "tags": [
          "Memories"
        ],
        "summary": "List memories",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/project"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Memory"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Memories"
        ],
        "summary": "Store a memory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "project": {
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memory": {
                      "$ref": "#/components/schemas/Memory"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/emails": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List emails",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/project"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "emails": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Email"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Ingest an email (raw or parsed)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "project": {
                    "type": "string"
                  },
                  "raw_email": {
                    "type": "string",
                    "description": "Raw RFC822 email content"
                  },
                  "parsed_email": {
                    "type": "object",
                    "description": "Mailparser-like JSON payload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "$ref": "#/components/schemas/Email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/urls": {
      "get": {
        "tags": [
          "URLs"
        ],
        "summary": "List URLs",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/project"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "urls": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Url"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "URLs"
        ],
        "summary": "Save a URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "title": {
                    "type": "string"
                  },
                  "project": {
                    "type": "string"
                  },
                  "crawl_enabled": {
                    "type": "boolean"
                  },
                  "screenshot_data_url": {
                    "type": "string",
                    "description": "Optional browser-captured screenshot data URL for authenticated or paywalled pages."
                  }
                },
                "required": [
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL already saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "$ref": "#/components/schemas/Url"
                    },
                    "duplicate": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "$ref": "#/components/schemas/Url"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/settings/byo-ai": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get BYO AI key status",
        "description": "Returns masked/configured status for hosted Pro account BYO AI API keys. Raw API keys are never returned.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "properties": {
                        "global": {
                          "$ref": "#/components/schemas/ByoAiScopeSettings"
                        },
                        "email": {
                          "$ref": "#/components/schemas/ByoAiScopeSettings"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Hosted Pro account admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update BYO AI API keys",
        "description": "Stores or clears hosted Pro account BYO AI API keys. Omitted fields are left unchanged; null or empty string clears a key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "global": {
                    "$ref": "#/components/schemas/ByoAiScopeUpdate"
                  },
                  "email": {
                    "$ref": "#/components/schemas/ByoAiScopeUpdate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "properties": {
                        "global": {
                          "$ref": "#/components/schemas/ByoAiScopeSettings"
                        },
                        "email": {
                          "$ref": "#/components/schemas/ByoAiScopeSettings"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid key payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Hosted Pro account admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/trash": {
      "get": {
        "tags": [
          "Trash"
        ],
        "summary": "List trashed items",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "notes",
                "memories",
                "urls",
                "emails"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trash"
        ],
        "summary": "Empty trash",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "deleted": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/graph": {
      "get": {
        "tags": [
          "Graph"
        ],
        "summary": "Get the knowledge graph data (nodes and edges)",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by project"
          },
          {
            "name": "include_tags",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "true"
            },
            "description": "Include tag-based edges"
          },
          {
            "name": "include_semantic",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "false"
            },
            "description": "Include semantic similarity edges"
          },
          {
            "name": "semantic_threshold",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 0.7
            },
            "description": "Semantic similarity threshold"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nodes": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "edges": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}